+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - Show file name on top of code panels.
 - Use the 'Esc' key to dismiss the code overlay window.
+- [WIP] Use 'Alt+1', 'Alt+2', 'Alt+3' to navigate between File explorer, reading area, and changes list. 
 
 ### Changed
 - Automatically scroll to the top of the code area after selecting a new file or revision.